requirements.txt
django==3.2.16
django-environ==0.11.2
django-cors-headers==4.2.0
djangorestframework==3.14.0
argon2-cffi==21.3.0
djoser==2.2.0
djangorestframework-simplejwt==5.2.2
social-auth-app-django==5.2.0
social.auth-core==4.4.1
PyJWT==2.5.0
Pillow==10.0.0
django-storages==1.13.2
django-ckeditor==6.7.0
psycopg2==2.9.7